feat: add error handling blueprint and refactor error handlers to improve code organization
